Break hands.h into yakus/scoring/util.
