+++ src/analysis/CMakeLists.txt
@@ -1,4 +1,12 @@
 target_sources(
   libmahjong
-  PRIVATE "handtree.cc" "possiblesets.cc" "hands.cc"
-  PUBLIC FILE_SET HEADERS BASE_DIRS "${libmahjong_SOURCE_DIR}/src" FILES "hands.h" "analysis.h")
+  PRIVATE "handtree.cc" "possiblesets.cc" "yakus.cc" "util.cc" "scoring.cc"
+  PUBLIC FILE_SET
+         HEADERS
+         BASE_DIRS
+         "${libmahjong_SOURCE_DIR}/src"
+         FILES
+         "yakus.h"
+         "analysis.h"
+         "util.h"
+         "scoring.h")
