+++ src/analysis/yakus.cc
@@ -1,457 +1,24 @@
-#include "hands.h"
+#include "analysis/yakus.h"
+#include "analysis/util.h"
 
 #include <algorithm>
 #include <array>
 #include <cstddef>
-#include <cstdint>
 #include <map>
-#include <memory>
 #include <set>
 #include <vector>
 
-#include "analysis/analysis.h"
 #include "statefunctions/stateutilities.h"
 #include "types/gamestate.h"
 #include "types/handnode.h"
 #include "types/meld.h"
 #include "types/pieces.h"
 #include "types/piecetype.h"
-#include "types/score.h"
 #include "types/statefunction.h"
 #include "types/walls.h"
 #include "types/winds.h"
 
 namespace mahjong {
-
-namespace {
-const int kMaxSingles = 15;
-
-int countSingles(const std::vector<Piece>& hand) {
-  auto root = breakdownHand(hand);
-  int min_singles = kMaxSingles;
-  for (const auto& branch : Node::AsBranchVectors(root.get())) {
-    int singles = 0;
-    for (const auto& node : branch) {
-      if (node->type() == Node::kSingle) {
-        singles++;
-      }
-    }
-    min_singles = std::min(singles, min_singles);
-  }
-  return min_singles;
-}
-}  // namespace
-
-int countPiece(const GameState& state, int player, Piece p) {
-  int count = 0;
-  for (const auto& piece : state.hands.at(player).live) {
-    if (piece == p) {
-      count++;
-    }
-  }
-  return count;
-}
-
-const std::vector<mahjong::yakuFunc> kYakuFunctions = {isFullyConcealedHand,
-                                                       isPinfu,
-                                                       isPureDoubleChi,
-                                                       isAllSimples,
-                                                       isMixedTripleChi,
-                                                       isPureStraight,
-                                                       isPrevalentWind,
-                                                       isSeatWind,
-                                                       isGreenDragon,
-                                                       isWhiteDragon,
-                                                       isRedDragon,
-                                                       isOutsideHand,
-                                                       isAfterAKan,
-                                                       isRobbingAKan,
-                                                       isBottomOfTheSea,
-                                                       isSevenPairs,
-                                                       isTriplePon,
-                                                       isThreeConcealedPons,
-                                                       isThreeKans,
-                                                       isAllPons,
-                                                       isHalfFlush,
-                                                       isLittleThreeDragons,
-                                                       isAllTerminalsAndHonors,
-                                                       isTerminalsInAllSets,
-                                                       isTwicePureDoubleChi,
-                                                       isBlessingOfMan,
-                                                       isFullFlush};
-
-const std::vector<yakuFunc> kYakumanFunctions = {
-    isThirteenOrphans, isNineGates,        isBlessingOfHeaven,
-    isBlessingOfEarth, isFourConcealedPon, isFourKans,
-    isAllGreen,        isAllTerminals,     isAllHonors,
-    isBigThreeDragons, isLittleFourWinds,  isBigFourWinds,
-    isMaxBranches};
-
-Score scoreHand(const GameState& state, int player) {
-  auto root = breakdownHand(state.hands.at(player).live);
-  Score s;
-  s.han = 0;
-  s.yakuman = 0;
-  s.fu = 0;
-  if (!root->IsComplete() && !isThirteenOrphans(state, player) &&
-      !isSevenPairs(state, player)) {
-    return s;
-  }
-  for (const auto& branch : Node::AsBranchVectors(root.get())) {
-    Score branchscore;
-    for (const auto& node : branch) {
-      if (node->type() == Node::kSingle) {
-        continue;
-      }
-    }
-
-    for (const auto& yaku_function : kYakuFunctions) {
-      branchscore.han += yaku_function(state, player, branch);
-    }
-    for (const auto& yaku_function : kYakumanFunctions) {
-      branchscore.yakuman += yaku_function(state, player, branch);
-    }
-    for (const auto& dora : state.walls.GetDoras()) {
-      for (const auto& p : state.hands.at(player).live) {
-        if (p == dora) {
-          branchscore.han++;
-        }
-      }
-      for (const auto& meld : state.hands.at(player).melds) {
-        if (meld.start == dora) {
-          branchscore.han++;
-        }
-        if (meld.type == Meld::kChi) {
-          if (meld.start + 1 == dora) {
-            branchscore.han++;
-          }
-          if (meld.start + 2 == dora) {
-            branchscore.han++;
-          }
-        }
-      }
-    }
-    branchscore.fu = getFu(state, player, branch);
-    if (getBasicPoints(branchscore) > getBasicPoints(s)) {
-      s = branchscore;
-    }
-  }
-
-  return s;
-}
-
-const int kLowHanBasicPoints = 2000;
-const int kManganBasicPoints = 2000;
-const int kManganMinHan = 5;
-const int kHanemanBasicPoints = 3000;
-const int kHanemanMinHan = 6;
-const int kBaimanBasicPoints = 4000;
-const int kBaimanMinHan = 8;
-const int kSanbaimanBasicPoints = 6000;
-const int kSanbaimanMinHan = 11;
-const int kYakumanBasicPoints = 8000;
-
-int getBasicPoints(Score s) {
-  if (s.yakuman > 0) {
-    return s.yakuman * kYakumanBasicPoints;
-  }
-  if (s.han >= kManganMinHan) {
-    if (s.han >= kSanbaimanMinHan) {
-      return kSanbaimanBasicPoints;
-    }
-    if (s.han >= kBaimanMinHan) {
-      return kBaimanBasicPoints;
-    }
-    if (s.han >= kHanemanMinHan) {
-      return kHanemanBasicPoints;
-    }
-    if (s.han >= kSanbaimanMinHan) {
-      return kSanbaimanBasicPoints;
-    }
-    if (s.han >= kManganMinHan) {
-      return kManganBasicPoints;
-    }
-  }
-  const int p = s.fu * (2 << (1 + s.han));
-  if (p > kLowHanBasicPoints) {
-    return kLowHanBasicPoints;
-  }
-  return p;
-}
-
-const int kSevenpairs = 25;
-const int kPinfuDiscard = 30;
-const int kPinfuSelfdraw = 20;
-const int kConcealedDiscard = 25;
-const int kOpenOrSelfdraw = 20;
-
-const int kSimplepon = 2;
-const int kTermHonorpon = 4;
-const int kSimplekan = 8;
-const int kTermHonorkan = 16;
-
-const int kConcealedMul = 2;
-const int kCsimplepon = kSimplepon * kConcealedMul;
-const int kCtermHonorpon = kTermHonorpon * kConcealedMul;
-const int kCsimplekan = kSimplekan * kConcealedMul;
-const int kCtermHonorkan = kTermHonorkan * kConcealedMul;
-
-const int kDragonseatprevalentwind = 2;
-const int kEdgeclosedpairwait = 2;
-const int kSelfdraw = 2;
-const int kOpenpinfu = 2;
-
-const int kLowedgewait = 3;
-const int kHighedgewait = 7;
-
-const int kFuRounding = 10;
-
-int getFu(const GameState& state, int player,
-          const std::vector<const mahjong::Node*>& branch) {
-  if (isSevenPairs(state, player, branch)) {
-    return kSevenpairs;
-  }
-  if (isPinfu(state, player, branch)) {
-    if (state.hasRonned.at(player)) {
-      return kPinfuDiscard;
-    }
-    return kPinfuSelfdraw;
-  }
-  int fu = 0;
-  if (!state.hands.at(player).open && state.hasRonned.at(player)) {
-    fu = kConcealedDiscard;
-  } else {
-    fu = kOpenOrSelfdraw;
-  }
-  const bool open = state.hands.at(player).open;
-  if (isOpenPinfu(state, player, branch)) {
-    fu += kOpenpinfu;
-  } else if (!state.hasRonned.at(player)) {
-    fu += kSelfdraw;
-  }
-  for (const auto& meld : state.hands.at(player).melds) {
-    if (meld.type == Meld::kKan) {
-      if (!meld.start.isHonor() && !meld.start.isTerminal()) {
-        fu += open ? kSimplekan : kCsimplekan;
-      } else {
-        fu += open ? kTermHonorkan : kCtermHonorkan;
-      }
-    }
-  }
-  for (const auto& node : branch) {
-    if (node->type() == Node::kPonSet) {
-      if (!node->start().isHonor() && !node->start().isTerminal()) {
-        fu += open ? kSimplepon : kCsimplepon;
-      } else {
-        fu += open ? kTermHonorpon : kCtermHonorpon;
-      }
-    }
-    if (open) {
-      continue;
-    }
-    if (node->type() == Node::kPair) {
-      if (node->start().isHonor()) {
-        if (node->start() == kGreenDragon || node->start() == kRedDragon ||
-            node->start() == kWhiteDragon ||
-            node->start() == (state.roundNum > 3 ? kSouthWind : kEastWind)) {
-          fu += kDragonseatprevalentwind;
-        }
-      }
-      if (node->start() == state.pendingPiece) {
-        fu += kEdgeclosedpairwait;
-      }
-    }
-    if (node->type() == Node::kChiSet) {
-      if (node->start() + 1 == state.pendingPiece) {
-        fu += kEdgeclosedpairwait;
-      }
-      if (node->start().getPieceNum() == 1 &&
-          state.pendingPiece.getPieceNum() == kLowedgewait &&
-          node->start().getSuit() == state.pendingPiece.getSuit()) {
-        fu += kEdgeclosedpairwait;
-      }
-      if (node->start().getPieceNum() == kHighedgewait &&
-          state.pendingPiece == node->start()) {
-        fu += kEdgeclosedpairwait;
-      }
-    }
-  }
-  if ((fu % kFuRounding) != 0) {
-    return fu;
-  }
-  return fu +
-         (kFuRounding - (fu % kFuRounding));  // rounding up to multiple of ten
-}
-
-bool isOpenPinfu(const GameState& state, int player,
-                 const std::vector<const mahjong::Node*>& branch) {
-  for (const auto& node : branch) {
-    if (node->type() == Node::kPonSet) {
-      return false;
-    }
-    if (node->type() == Node::kPair) {
-      if (node->start() == kRedDragon || node->start() == kWhiteDragon ||
-          node->start() == kGreenDragon) {
-        return false;
-      }
-      if (node->start() == kSouthWind && state.roundNum > 3) {
-        return false;
-      }
-      if (node->start() == kEastWind && state.roundNum < 4) {
-        return false;
-      }
-    }
-  }
-  for (const auto& meld : state.hands.at(player).melds) {
-    if (meld.type > Meld::kChi) {
-      return false;
-    }
-  }
-  auto waits = isInTenpai(state.hands.at(player).live, /*allWaits=*/true);
-  return waits.size() != 1;
-}
-
-bool isComplete(const GameState& state, int player) {
-  auto root = breakdownHand(state.hands.at(player).live);
-  if (!root->IsComplete() && !isThirteenOrphans(state, player)) {
-    return false;
-  }
-  for (const auto& branch : Node::AsBranchVectors(root.get())) {
-    if (std::any_of(branch.begin(), branch.end(), [](const auto& node) {
-          return node->type() == Node::kSingle;
-        })) {
-      continue;
-    }
-
-    for (const auto& yaku_function : kYakuFunctions) {
-      if (yaku_function(state, player, branch) > 0) {
-        return true;
-      }
-    }
-    for (const auto& yaku_function : kYakumanFunctions) {
-      if (yaku_function(state, player, branch) > 0) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-const std::vector<Piece> kPieceSet{
-    kOneBamboo,    kTwoBamboo,      kThreeBamboo,    kFourBamboo,
-    kFiveBamboo,   kSixBamboo,      kSevenBamboo,    kEightBamboo,
-    kNineBamboo,   kOnePin,         kTwoPin,         kThreePin,
-    kFourPin,      kFivePin,        kSixPin,         kSevenPin,
-    kEightPin,     kNinePin,        kOneCharacter,   kNineCharacter,
-    kTwoCharacter, kThreeCharacter, kFourCharacter,  kFiveCharacter,
-    kSixCharacter, kSevenCharacter, kEightCharacter, kWhiteDragon,
-    kGreenDragon,  kRedDragon,      kEastWind,       kSouthWind,
-    kNorthWind,    kWestWind};
-
-// TODO(#20): This is an extremely inefficient algorithm but it's probably good enough for
-// the frequency it needs to be ran
-// will revisit if necessary
-// assumption is 14 piece hand
-std::vector<Piece> isInTenpai13Pieces(std::vector<Piece> hand, bool allWaits) {
-  const int min_singles = countSingles(hand);
-  // These numbers were found by looking at a lot of handtrees and their single count
-  if (min_singles > 5 || min_singles == 3 || min_singles == 0) {
-    return {};
-  }
-  std::array<int8_t, Piece::kPiecesize> counts = {};
-  std::vector<Piece> waits;
-  for (const auto& p : hand) {
-    counts.at(p.toUint8_t())++;
-  }
-  for (const auto& p : kPieceSet) {
-    if (counts.at(p.toUint8_t()) == 4) {
-      continue;
-    }
-    hand.push_back(p);
-    auto root = breakdownHand(hand);
-    if (root->IsComplete()) {
-      waits.push_back(p);
-      if (!allWaits) {
-        return waits;
-      }
-    }
-
-    hand.pop_back();
-  }
-  return waits;
-}
-
-const int kPiecesinahand = 14;
-
-std::vector<Piece> isInTenpai(std::vector<Piece> hand, bool allWaits) {
-  if (hand.empty()) {
-    return {};
-  }
-  const int min_singles = countSingles(hand);
-  // These numbers are the same as above except one more piece means 1 higher on the single count
-  if (min_singles > 6 || min_singles == 1 || min_singles == 4 ||
-      min_singles == 0) {
-    return {};
-  }
-  std::array<bool, Piece::kPiecesize> removedbefore = {};
-  std::vector<Piece> waits;
-  for (int i = 0; i < kPiecesinahand; i++) {
-    const Piece removed = hand.front();
-    hand.erase(hand.begin());
-    if (removedbefore.at(removed.toUint8_t())) {
-      hand.push_back(removed);
-      continue;
-    }
-    removedbefore.at(removed.toUint8_t()) = true;
-    std::vector<Piece> tempwaits = isInTenpai13Pieces(hand, allWaits);
-    if (!tempwaits.empty()) {
-      if (!allWaits) {
-        return tempwaits;
-      }
-      waits.insert(waits.begin(), tempwaits.begin(), tempwaits.end());
-    }
-    hand.push_back(removed);
-  }
-  return waits;
-}
-
-// see above comment
-std::vector<Piece> getRiichiDiscard(std::vector<Piece> hand) {
-  if (hand.empty()) {
-    return {};
-  }
-  std::array<int8_t, Piece::kPiecesize> counts = {};
-  std::array<bool, Piece::kPiecesize> removedbefore = {};
-  std::vector<Piece> remove_me;
-  for (const auto& p : hand) {
-    counts.at(p.toUint8_t())++;
-  }
-  for (int i = 0; i < kPiecesinahand; i++) {
-    const Piece removed = hand.front();
-    hand.erase(hand.begin());
-    if (removedbefore.at(removed.toUint8_t())) {
-      hand.push_back(removed);
-      continue;
-    }
-    removedbefore.at(removed.toUint8_t()) = true;
-    for (const auto& p : kPieceSet) {
-      if (counts.at(p.toUint8_t()) == 4 || p == removed) {
-        continue;
-      }
-      hand.push_back(p);
-      auto root = breakdownHand(hand);
-      if (root->IsComplete()) {
-        remove_me.push_back(removed);
-      }
-
-      hand.pop_back();
-    }
-    hand.push_back(removed);
-  }
-  return remove_me;
-}
-
 bool isRiichi(const GameState& state, int player,
               const std::vector<const mahjong::Node*>& /*branch*/) {
   return state.hands.at(player).riichi && !state.hands.at(player).open;
@@ -466,8 +33,7 @@ bool isDoubleRiichi(const GameState& state, int player,
 bool isIppatsu(const GameState& state, int player,
                const std::vector<const mahjong::Node*>& branch) {
   return isRiichi(state, player, branch) &&
-         (state.turnNum - state.hands.at(player).riichiRound <=
-              4 &&
+         (state.turnNum - state.hands.at(player).riichiRound <= 4 &&
           state.lastCall < state.hands.at(player).riichiRound);
 }
 
@@ -552,12 +118,12 @@ bool isAllSimples(const GameState& state, int player,
   return true;
 }
 
-const int kPiecesinasuit = 9;
 bool isMixedTripleChi(const GameState& state, int player,
                       const std::vector<const mahjong::Node*>& branch) {
-  std::array<bool, kPiecesinasuit> bamboo_chi = {};
-  std::array<bool, kPiecesinasuit> char_chi = {};
-  std::array<bool, kPiecesinasuit> pin_chi = {};
+  const int k_piecesinasuit = 9;
+  std::array<bool, k_piecesinasuit> bamboo_chi = {};
+  std::array<bool, k_piecesinasuit> char_chi = {};
+  std::array<bool, k_piecesinasuit> pin_chi = {};
   for (const auto& node : branch) {
     if (node->type() == Node::kChiSet) {
       if (node->start().getSuit() == Piece::Type::kBambooSuit) {
@@ -584,7 +150,7 @@ bool isMixedTripleChi(const GameState& state, int player,
       }
     }
   }
-  for (int i = 0; i < kPiecesinasuit; i++) {
+  for (int i = 0; i < k_piecesinasuit; i++) {
     if (bamboo_chi.at(i) && char_chi.at(i) && pin_chi.at(i)) {
       return true;
     }
@@ -592,22 +158,22 @@ bool isMixedTripleChi(const GameState& state, int player,
   return false;
 }
 
-const int kFirstchistart = 1;
-const int kSecondchistart = 4;
-const int kThirdchistart = 7;
 bool isPureStraight(const GameState& state, int player,
                     const std::vector<const mahjong::Node*>& branch) {
+  const int k_firstchistart = 1;
+  const int k_secondchistart = 4;
+  const int k_thirdchistart = 7;
   std::array<bool, 3> bamboo_chi = {};
   std::array<bool, 3> char_chi = {};
   std::array<bool, 3> pin_chi = {};
   for (const auto& node : branch) {
     if (node->type() == Node::kChiSet) {
       int ind = 0;
-      if (node->start().getPieceNum() == kFirstchistart) {
+      if (node->start().getPieceNum() == k_firstchistart) {
         ind = 0;
-      } else if (node->start().getPieceNum() == kSecondchistart) {
+      } else if (node->start().getPieceNum() == k_secondchistart) {
         ind = 1;
-      } else if (node->start().getPieceNum() == kThirdchistart) {
+      } else if (node->start().getPieceNum() == k_thirdchistart) {
         ind = 2;
       } else {
         continue;
@@ -626,11 +192,11 @@ bool isPureStraight(const GameState& state, int player,
   for (const auto& meld : state.hands.at(player).melds) {
     if (meld.type == Meld::kChi) {
       int ind = 0;
-      if (meld.start.getPieceNum() == kFirstchistart) {
+      if (meld.start.getPieceNum() == k_firstchistart) {
         ind = 0;
-      } else if (meld.start.getPieceNum() == kSecondchistart) {
+      } else if (meld.start.getPieceNum() == k_secondchistart) {
         ind = 1;
-      } else if (meld.start.getPieceNum() == kThirdchistart) {
+      } else if (meld.start.getPieceNum() == k_thirdchistart) {
         ind = 2;
       } else {
         continue;
@@ -787,9 +353,9 @@ bool isSevenPairs(const GameState& state, int player,
 
 bool isTriplePon(const GameState& state, int player,
                  const std::vector<const mahjong::Node*>& branch) {
-  std::array<bool, kPiecesinasuit> bamboo_pon = {};
-  std::array<bool, kPiecesinasuit> char_pon = {};
-  std::array<bool, kPiecesinasuit> pin_pon = {};
+  std::array<bool, 9> bamboo_pon = {};
+  std::array<bool, 9> char_pon = {};
+  std::array<bool, 9> pin_pon = {};
   for (const auto& node : branch) {
     if (node->type() == Node::kPonSet) {
       if (node->start().getSuit() == Piece::Type::kBambooSuit) {
@@ -816,7 +382,7 @@ bool isTriplePon(const GameState& state, int player,
       }
     }
   }
-  for (int i = 0; i < kPiecesinasuit; i++) {
+  for (int i = 0; i < 9; i++) {
     if (bamboo_pon.at(i) && char_pon.at(i) && pin_pon.at(i)) {
       return true;
     }
