+++ src/api/gamestate.cc
@@ -2,19 +2,12 @@
 
 #include <algorithm>
 #include <memory>
-#include <array>
-#include <map>
-#include <string>
+#include <new>
 #include <vector>
-
-#include "api/types.h"
-#include "controllers/controllermanager.h"
 #include "statefunctions/statecontroller.h"
 #include "types/gamestate.h"
 #include "types/settings.h"
 #include "types/statefunction.h"
-#include "types/hand.h"
-#include "types/piecetype.h"
 
 namespace api {
