+++ src/statefunctions/gamestates/roundend.cc
@@ -53,12 +53,12 @@ std::unique_ptr<GameState> RoundEnd(std::unique_ptr<GameState> state) {
     state->players.at(i).points += state->scores.at(i);
   }
   state->scores = {};
-  
+
   // TODO (#14): for now naively increment the round number
   state->roundNum++;
 
   // TODO (#16): East wind only for now
-  if (state->roundNum > last_round) { 
+  if (state->roundNum > last_round) {
     state->nextState = StateFunctionType::kGameEnd;
   } else {
     state->nextState = StateFunctionType::kRoundStart;
