+++ tests/yakus/afterkan.test.cc
@@ -4,7 +4,8 @@
 #include <string>
 
 #include "analysis/analysis.h"
-#include "analysis/hands.h"
+#include "analysis/yakus.h"
+
 #include "types/gamestate.h"
 #include "types/hand.h"
 #include "types/handnode.h"
