+++ tests/yakus/allpon.test.cc
@@ -5,7 +5,8 @@
 #include <vector>
 
 #include "analysis/analysis.h"
-#include "analysis/hands.h"
+#include "analysis/yakus.h"
+
 #include "types/gamestate.h"
 #include "types/hand.h"
 #include "types/handnode.h"
